#39 | [@shrijayan] refactor: Update invitation handling by removing transaction context and simplifying member insertion logic
